ci: drop testing of jdk8 on macos (#154)
**circleci: configuring for Apple silicon**

I could not find a combination of an m1 `resource_class` and `xcode`
that did not produce a "not available for your project" error from
CircleCI. So I left out `resource_class` for now. Perhaps all will be
good after Apple Silicon is the only supportted macos executor on
June 29, 2024. I'll come back and check on this an open a new issue if
necessary.

**circleci: jdk: infer os and architecture from host**

JDK architecture was hardcoded to x64 but this is not appropriate for our new
Apple silicone hardware. We now infer architecture from host.

Figured might as well infer os from host as well.

Adjusted install dir name to include os and architecture to support
testability and a potential future where one might, for example, install
x64 on Apple Silicon hardware and use Rosetta. The latter would require
a tweak to allow architecture to be selected by the user. We'll allow
for that in the future should we need it.

Closes #153